fix(review): address OCR review threads on stack/21 (PR 93 sweep)
Restore core behavior: QueryPlanner final-row SeedSubstitutesOnly application, QueryPlanner.h header style, BitUtils forEachSetBit (fixes BitUtilsTest build). Correctness: zero-copy MORE short-send check, MSG_NOSIGNAL on async sends, zero-send guard, pool overflow guards; io_uring submit retry, buffer-offset validation, per-batch completion accounting; dispatcher NaN/INF, date datatype, int datatype, tag range; RlePrefix term-keyed cache + overflow pre-checks; MonoSer Undefined fallback + noexcept; FastExport integer reservation + typed-literal escaping; HttpChunk capacity check; Prefetching bounds + wraparound; ExportIds/PolyVocab builder-result protocol; SIMD scan/escape benchmark methodology. Hygiene: workflow timeouts, missing includes, designated-init to positional, ceil-cast clamps, test socket RAII/timeout/join guards. clang-format-18 clean.

+++ benchmark/FsstScratchBufferBenchmark.cpp
@@ -91,13 +91,14 @@ class FsstScratchBufferBenchmark : public BenchmarkInterface {
         "Three-stage FSST scratch-buffer strategies (5,000 words)");
     const char* selectedStrategyEnv = std::getenv("FSST_SCRATCH_ONLY");
     const size_t selectedStrategy =
-        selectedStrategyEnv == nullptr
-            ? 3
-            : std::strtoul(selectedStrategyEnv, nullptr, 10);
+        selectedStrategyEnv == nullptr ? 3
+                                       : static_cast<size_t>(std::strtoul(
+                                             selectedStrategyEnv, nullptr, 10));
     const char* repetitionsEnv = std::getenv("FSST_SCRATCH_INNER_REPETITIONS");
-    const size_t repetitions = repetitionsEnv == nullptr
-                                   ? 1
-                                   : std::strtoul(repetitionsEnv, nullptr, 10);
+    const size_t repetitions =
+        repetitionsEnv == nullptr
+            ? 1
+            : static_cast<size_t>(std::strtoul(repetitionsEnv, nullptr, 10));
     AD_CONTRACT_CHECK(selectedStrategy <= 3);
     AD_CONTRACT_CHECK(repetitions > 0);
